refactor(ui): split menu wrappers for parity
Align the UI menu implementation with components behavior and styling, while keeping the public Menu API composed from focused wrappers.
