+++ packages/ui/src/menu/index.ts
@@ -1,2 +1,41 @@
-export { Menu } from './menu';
+import { CheckboxItem } from './checkbox-item';
+import { CheckboxItemIndicator } from './checkbox-item-indicator';
+import { Group } from './group';
+import { GroupLabel } from './group-label';
+import { Item } from './item';
+import { ItemHelpText } from './item-help-text';
+import { ItemLabel } from './item-label';
+import { Popup } from './popup';
+import { Portal } from './portal';
+import { Positioner } from './positioner';
+import { RadioGroup } from './radio-group';
+import { RadioItem } from './radio-item';
+import { RadioItemIndicator } from './radio-item-indicator';
+import { Root } from './root';
+import { Separator } from './separator';
+import { SubmenuRoot } from './submenu-root';
+import { SubmenuTriggerItem } from './submenu-trigger-item';
+import { Trigger } from './trigger';
+
+export const Menu = {
+	Root,
+	Trigger,
+	Portal,
+	Positioner,
+	Popup,
+	Item,
+	ItemLabel,
+	ItemHelpText,
+	CheckboxItem,
+	CheckboxItemIndicator,
+	RadioItem,
+	RadioItemIndicator,
+	Separator,
+	Group,
+	GroupLabel,
+	RadioGroup,
+	SubmenuRoot,
+	SubmenuTriggerItem,
+};
+
 export * from './types';
