+++ packages/ui/src/menu/root.tsx
@@ -1,0 +1,8 @@
+import { Menu as BaseMenu } from '@base-ui/react/menu';
+import type { MenuRootProps } from './types';
+
+function Root( props: MenuRootProps ) {
+	return <BaseMenu.Root { ...props } />;
+}
+
+export { Root };
