+++ packages/ui/src/menu/submenu-root.tsx
@@ -1,0 +1,9 @@
+import { Menu as BaseMenu } from '@base-ui/react/menu';
+import type { MenuSubmenuRootProps } from './types';
+
+function SubmenuRoot( props: MenuSubmenuRootProps ) {
+	return <BaseMenu.SubmenuRoot { ...props } />;
+}
+SubmenuRoot.displayName = 'Menu.SubmenuRoot';
+
+export { SubmenuRoot };
